requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0
jinja2>=3.1.0
dnspython>=2.2.0
python-whois>=0.7.0
python-nmap>=0.7.1
requests-cache>=0.9.0
scapy>=2.4.5
colorama>=0.4.4
rich>=12.0.0
typer>=0.7.0